api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
polars==0.20.0
duckdb
scikit-learn==1.3.2
mlflow==2.9.2
numpy==1.24.4
pytest==7.4.3
